refactor : Canvas Entity owner 추가와 패키지 이동 (#12)
